caos: Rely on persistent storage backend for consistency checking
The only place that can guarantee that certain data conforms to the
consistency and correctness criteria is the persistent storage backend
itself.  Ergo, from now on the handling strategy for consistency criteria
that involves more than one immediate entity is to rely on the storage
backend ACID properties and re-interpret and re-package exceptions to
provide proper context.

One effect of this is that consistency violations will invalidate
transactions to the nearest savepoint, so sufficient care should be
taken with regards to the context in which data commit operations are
performed.