caos: Use the topmost source in index pointer expressions
Index pointer expressions used to infer the source of the pointer by
calling get_pointer_origin().  However, the default behaviour of it is
to return the closest source that (re-)defines the specified link, not
the topmost one.  This leads to an inconsistency when inhertied indexes
are interpreted by the PostgreSQL backend, since there is no way to
reliably tell which concept was returned by get_pointer_origin().  Using
the topmost source does not have that issue.