caos: Unify, generalize and improve constraint handling
Atom constraints, link constraints and link property constraints are now
parts of a more general data consistency enforcement framework.

Handling of pointer unique constraints has been much improved to the
point where they work properly in different inheritance scenarios.

Additionally, a new mechanism of abstract constraints has been added.
In this context, "abstract" means that a constraint takes effect on the
children of the source that defines the constraint and not the source
itself.