caos: Add support for expressional pointer defaults
In addition to literal values it is now possible to use CaosQL queries
to specify pointer defaults.  This effectively enables defaults
functionality for non-atomic pointers.

Defaults are now tried until all options have been exhausted or a
non-empty result is obtained, i. e., it is possible to specify multiple
queries in link default section and the first one yielding a non-empty
value will be accepted.

There are two main restrictions for queries in defaults:

  * Return value type must match the type of the target
  * For non-atomic links the link mapping must be either "*1" or "**"
  * Query result must match pointer cardinality, so for "11" and "*1"
    mappings a single result must be returned.