caos: Add "anchor" functionality to selector expressions
This implements an analogue to CaosQL's {anchor} notion where a path can
be explicitly marked by an anchor differentiating if from the other
paths along the same graph trail which prevents the paths from merging.

To anchor some path one has to call the detach() method optionally
providing the anchor name.

Example:

    C1 = schema.Country.detach()
    C2 = schema.Country.detach()

    qry = C2.query(country=C2.name, ethnicity=C2.people.ethnicities.name) \
            .filter(C1.people.ethnicities == C2.people.ethnicities,
                    C1.id != C2.id, C1.name == 'France') \
            .order(C2.name).order(C2.people.ethnicities.name)

Here, C1 and C2 are different paths forming a join with a custom condition.
The above expression is equivalent to the following CaosQL query:

    SELECT
        C2.name AS country,
        C2.people.ethnicities.name AS ethnicity
    WHERE
        Country{C1}.people.ethnicities = Country{C2}.people.ethnicities
        AND C1 != C2 AND C1.name = 'France'
    ORDER BY
        C2.name, C2.people.ethnicities.name