caos: Adjust list_sum() to use constant arrays whenever possible
When an argument to list_sum() is a sequence of constants it is much
more efficient to represent it as a constant of an array type rather
than as a sequence of constants.  Furthermore, the sequence must contain
elements of the same type, and this commit adds a check that enforces
this requirement.