caos: Add support for computable links
Computable links appear and behave the same as regular read-only links,
except that their value does not come from the storage, but from a
result of an evaluation of a specific expression defined in the data
schema. The expression can be arbitrary, although, its return value must
be either a single scalar or a set of entities.

For example, an Order concept could have "subtotal" and "sales_tax_rate"
as its links and then define the "total" as a computable:

concepts:
    Order:
        links:
            subtotal:
                money_t:
                    title: "Order Subtotal"

            sales_tax_rate:
                rate_t:
                    title: "Sales Tax Rate"

        computables:
            total: self.subtotal + self.subtotal * self.sales_tax_rate

This commit adds the necessary infrastructure and adds support for basic
computables.  Support for computables in queries will be added in further
commits.