caos: Make session state consistent wrt to transaction status
This makes session behaviour more consistent with respect to transaction
state.  From now on the session is automatically sync()'ed on each
transaction boundary.  Changes recorded during a transaction that was
rolled back are reverted: new entities are detached from the session and
turned into "zombie" entities to prevent their further use, modified
entities get their data restored to the pre-transaction state and
deleted entitites are resurrected.

As such, the Session is in a consistent state and can be safely used
after a transaction rollback.