utils.lang.yaml.validator: Make it possible to import schemas into schemas
It is now possible to import other document schemas into other schemas
and reference them inside the importing schema.  This removes the
"single-file-schema" limitation and makes it possible to modularize and
re-use schemas.

The actual change is trivial and makes me wonder why we didn't do this
before.  The rest is validator tests reshuffling to make test schemas
imported as modules instead of loading them directly.