caos: Implement support for lazy loading of atomic links
Generally, the amount of data in atomic links is fairly small and the
most efficient way is to load all atomic links in a bunch -- so called
"eager" loading.  This is the default for atomic links.

In some cases, however, unnecessary loading of a certain atomic link can
be costly, mostly when a lot of data has to be transferred from the
backend.  So, introduce a mechanism to control link loading behaviour
with the "loading" attribute.  Currently, two values are allowed:
'eager' and 'lazy'.  If set to 'lazy', the link will never be loaded in
batch, only from direct dereference.

This is intended to be a generic mechanism to control link loading.
This commit, however, addresses only atomic links.