caos: Adjust handling and interpretation of link target types
When a link points to an atom, it is deemed to be 'atomic' and
'non-atomic' otherwise.  As only concrete links can have specific
targets, atomicity is a property inherent to concrete links.

Previously, this property propagated to the parent generic link and if
other concrete links stemming from this generic link had mismatching
target type, an exception had risen.

There are numerous problems with this approach.  First, it restricts
possibly unrelated links that happen to have the same name to have the
same target type.  Second, for abstract generic links the property of
atomicity is not determined at all.

So, from now on only concrete links have determined atomicity, generic
links have it undefined and any attempt to reference it will cause an
exception.  All call sites have been fixed not to rely on the value of
atomic() for generic links.