utils: Implement clear_exc_info(), set_exc_info() helpers
It turned out that greenlets do not manage exception context properly
when switching, so we need to do that ourselves.  As it is not possible
(or very cumbersome) to modify exc_info on Python level, introduce
semantix.utils.cutils C module and place clear_exc_info() and
set_exc_info() methods to correspondingly clear and set exception
context.