utils.functional.Decorator: Plug a ginormous memory leak
The Decorator machine on which many Semantix decorators are based
contained a memory leak: although the decorator cache uses a
WeakKeyDictionary, the functional.partial object which was decorated and
cached contained a direct reference to an instance on which the
decorated method was called.  The proper fix is to use weakref.