utils.functional: Add 'callable' function and start using it across the framework
Check whether the object is callable by calling hasattr(obj, '__call__')
is generally incorrect.  The main reason is that the 'hasattr'
function calls 'getattr' behind the scenes and so, for instance,
if __call__ is implemented via the descriptor protocol with some
code in its __get__ then this code will be executed.

Current implementation is borrowed from python collections module and
examines just the type's structure.

Start using the function by importing it from the functional module
directly. Later, when python 3.2 is released the current implementation
will be dropped.