caos: Always use symbolic names for query constants
Previously query arguments generated from Python expression constants
were given numeric indexes.  This made it hard to amend the expression
and the constants list, and was incompatible with the symbolic naming of
arguments in CaosQL queries.  So just make sure that query constants are
always assigned an unambiguous symbolic name.