caos.tree: Band-aid solution for set meta filter merging
IS and IS NOT operators are called 'meta filters', because they operate
on the meta level, rather than on data level.  Unfortunately, the
current coding handles them rather poorly by applying the filter very
early in the process, so correct propagation of the filter during path
merging is difficult.  This is a band-aid solution for the problem that
makes simple queries work.