caos.backends.pgsql: Fix cardinality estimation for inbound link paths
The cardinality estimation for a path did not work properly for paths
with reversed links, due to a fallacy in thinking that link direction is
the same as query vector direction.  In reality, those are completely
different things, and query vector is always considered from root to
leaves regardless of link directions in any of path components.