caos.backends.pgsql: Introduce backend format version tracking and upgrade
The format of various backend structures and data quite naturally
changes over time.  Ergo, there has to be a clear mechanism to upgrade
existing backend instances to the most current version of backend
format.

An example of that is the recent introduction of feature tracking.  This
commit retroactively records features present in backends initialized
before the functionality was added.