caos: Fix prototype schema index
The original implementation of prototype schema index was quite broken
because it relied on an incorrect __hash__() override for prototype
objects.  While in some cases it is useful to have sets of prototypes
hashed by prototype name, it is a bad idea in general, because
overridden __hash__() also requires corresponding __eq__() which would
lead to all sorts of unintended consequences.

So, instead of overriding __hash__() use explicit dicts with name as a
key and when set semantics is required, employ ExtendedSet with explicit
key function.