utils.lang.python: Do not use inspect.getmodule()
inspect.getmodule() carelessly iterates over sys.modules to build a
filename to module mapping.  Since the iteration is done directly on
sys.modules and attempts are made to access module attributes, a lazy
module reloading may be triggered, which causes sys.modules to change
and the loop to crash with RuntimeError.  So, stop using
inspect.getmodule() and switch to lang.Importer.get_module_by_filename().