caos: Always use specialized pointer prototypes for link subclasses
For every created link class corresponding to a (source, link,
target) triple, a proper corresponding link prototype should be used.
If such prototype was not statically defined in the schema, fallback to
using a generic prototype.  The latter should only happen for
dynamically constructed paths and computables, and a fully correct
solution would be to dynamically construct a prototype for each such
link class, however, that would be covered in subsequent commits.