sx: Add support for reporting code coverage via --coverage
sx learned to track and report code execution coverage with the help of
"coverage" module.  The "--coverage=<module>" argument controls whether
to enable coverage statistics and which module(s) to include.

For best results coverage tracing needs to be started as early as
possible, thus semantix initialization sequence was refactored so that
--profile and --coverage arguments are parsed before most of semantix
modules are even loaded.

--profile and --coverage are mutually exclusive.