caos: Implement support for volatile queries
A query is considered volatile if the values of its arguments have
drastic effect on the efficiency of the query plan.  As such, backend query
plans for such queries cannot be cached and a new plan has to be
generated for each invokation.

This mostly stems from the backend database behaviour --- there is
always a tradeoff between the savings on query planning and the
potential for suboptimal plan being generated.  It is now possible to
control the planning behaviour by calling

    selector.mark(volatile=True)

This will instruct the query engine not to attempt to use generic plan
caching for this query and generate a new plan for every argument set.
A side effect of this is that argument references will be substituted
with string representation of argument values, which means that each
argument must have a string representation compatible with database
backend.  Consequently, every datatype will need to also have to define
string IO routines to complement the existing binary IO routines.  That
is a material for future commits --- quote(str(arg)) will suffice for now.