caos: Get rid of linkset indirection
Linkset indirection is an unfortunate design artifact created originally
to support multi-target concept links, like the following:

    Source:
    	link: [Target1, Target2]

Where Target1 and Target2 are normally unrelated.  It may be argued that
a common parent could be used, but in practice it may not always be
possible to alter the target bases list to introduce that common parent.
Thus, linksets were introduced as an indirection, so instead of normal

    Source - link - Target

there would be
                      /- link' - Target1
    Source - linkset -
                      \- link'' - Target2

As one may imagine, this complicates things quite a bit, littering the
code all over with special handling of linksets.  Additionally, because
there is no single target, class attribute dereference would yield not
the target class as is normally expected, but another shim object --
NodeSet -- which represents a set of potential classes.  Referring to
NodeSet attributes would yield new NodeSet objects and so on.

Another major problem is that current implementation handles mixed
target sets incorrectly, i. e. when entity link targets actually contain
entities of different classes, so only homogeneous sets are handled
reliably, which defeats the whole purpose of multitarget links.

As it turns out, the original problem, that creating a common parent
and pointing to it instead may not be possible can be solved by
introducing a special type of concepts called "virtual concepts".
Whenever a multitarget condition occurs -- whether through explicit
definition in the schema, or as a result of pointer resolution -- a
concept prototype will be created on the fly having the targets as its
direct children.  This concept is marked as virtual.  Implicit virtual
concepts (i.e those not mandated directly by the schema definition) are
not even registered normally in the persistent storage backend.

This rework brings significant simplifications across Caos, netting 100
lines less and one less failing test.  The cleanup is not throughout,
notably, most linkset iteration loops remain -- I chose to minimize code
churn.