utils.lang.yaml: Sanitize YAML serialization API
Current implementation of YAML serialization framework has a fairly
careless API.  First, the protocol methods are called 'construct' and
'represent' --- too generic and non-obvious.  Second, a reference to
raw object data is stored directly in the 'data' attribute of the
object.

Fix the issue by:

* Renaming construct() to __sx_setstate__() and represent() to
  __sx_getstate__(), making it obvious that those are magic methods
  and are related to serialization.

* Halting the practice of writing raw data into object and passing
  it as an argument to __sx_setstate__() instead.

There's another bit YAML serializer pollutes the object with: source
context.  That will be dealt with in subsequent commits.