caos: Make it possible to Turn CaosQL expressions into selectors
This adds a new method: Session.selector(), which takes textual
representation of a query and returns either an entity Selector or a
QuerySelector, depending on the expression.  The resulting selector then
behaves identically to the one obtained via Python expressions for the
same query, i.e. can be augmented with calls to filter(), order() etc.
Naturally, CaosQL expressions can be used as subqueries via this
interface.

To support parametrized selectors, which is practically all
CaosQL-originated selectors, Selector class gained a method to apply
argument values to the selector: apply(**kwargs).  __call__() cannot be
used, because entity selectors are classes.

This work completes the unification of Caos selector interface and makes
all other interfaces deprecated and obsolete.