utils.cache: Implement basic cache infrastructure.
To use it - derive a class from utils.cache.Bucket, instantiate it and
use as a regular dictionary object.

Buckets accept two optional arguments - 'bucket_id', and 'parent'.
The first one is needed to differenciate buckets instantiated from the
same bucket class, the second one is to introduce dependencies.  When
parent bucket is reset, all its children are reset too.

Another level of cache hierarchy lays on a class level.  If the 'reset'
method is called on a bucket class, then all instances of that class and
derived classes will be reset.