caos: Add support for custom processing to schema deltas
It is now possible to provide custom logic for schema deltas as Python
code.  Every delta command will now execute code stored in its
"preprocess" attribute before the delta is applied to a schema.
Similarly, code in "postprocess" attribute will be executed after the
delta is applied to a schema.  The code itself must contain a callable
"main" which takes session as its only argument.  There are no
restrictions as to what the code can do.

`sx caos meta add` learned to accept preprocess and postprocess code via
`--pre` and `--post` arguments correspondingly, which each should contain
a name of the file to copy the code from.

Besides custom schema processing, this facility can be used for data
migrations.  When either `--pre` or `--post` argument is specified,
`sx caos meta add` will record a delta regardless of whether there are
changes to the schema.