caos: Fix IS [NOT] <type> operator in queries
Current implementation of "IS [NOT] <type>" operator is broken to the
point of uselessness: it attempts to do all the work early on the Caos
tree level by constructing an elaborate path combination based on the
arguments.  Besides considerably complicating the query tree, this also
generates a new query tree for each argument combination, even if
constant.

The correct way of doing this is, of course, to treat the IS [NOT]
operator just like any other binop and let the backend transformer deal
with it.  For database backend the task of type filtering neatly
transforms into a local expression on the type id column.  To avoid
overhead of scanning tables high in the hierarchy tree we can enable
constraint exclusion based on the type id column, although this commit
does not address that.