caos.backends.pgsql: Improve handling of subqueries
Current coding puts almost all subqueries, regardless of their position
in the source query, into the FROM SQL clause.  While semantically
valid, this may produce substantially non-optimal SQL queries.  Thus,
make it so that subqueries originally in the selector expression are
placed into the SELECT list.  This prohibits multi-attribute subqueries
in selector, so xfail the corresponding test.  It's not clear whether
this functionality is necessary at all, though, given that it is now
possible to fetch and aggregate arbitrary tuples.