caos.caosql: Fix ``<expr> IS [NOT] type`` in aggregates (and elsewhere)
In many contexts a reference to a concept node is interpreted as an
identity reference, but the typecheck operator is one of the contexts,
where this is not the case.  A better fix is, of course, to push the
entityref_to_idref transformation down the pipeline, but that requires
somewhat more work.