caos: Push handling of children-only pointers to backend transformer
Caos semantics normally allow references to pointers which are not
defined in a source, but are defined in one or more of its descendants.
This semantics is not directly compatible with that of the backend SQL
database, where a reference to a table column requires its existence in
the parent table.  For this reason, CaosQL and Python query transformers
resorted to various hackery, including replacing the source class, which
in turn resulted in invalid query tree merges and other badness.

The right approach is to push the handling of children-only pointers
down to the backend transformer.  For SQL backend the workaround for
"missing" columns is to LEFT JOIN child tables actually containing the
column by id and use them for references.