utils.lang.js: Add new concept - module hooks.
When a non-javascript module is imported in a context of a javascript
one, and module hooks were registered on the loader, we now try to
iterate through all attributes of the former module and see if a hook
can convert it to a resource reusable by javascript.  For example, this
may be used to add an ability of importing YAML-defind grids and forms
directly in javascript.