utils.lang: Make it possible to override module class late
In certain cases a loader is only able to determine the module class
requirement late in the loading process.  For example, a YAML module
schema may require the loaded module be of a certain type.

So, if the loaded module contains the '__sx_moduleclass__' attribute, it
will be used to reconstruct the module.