caos: Allow non-singular links to atoms
Since inception, there's been a limitation that atoms could only be
referenced by a link with one-to-one mapping.  This stems from the
backend implementation, since atoms are stored in a single table as
columns.  This limitation led to the necessity of creating intermediate
wrapper concepts when multiple values needed to be linked, even if the
value is an atom.

This rework removes the singular-only limitation for atomic links by
making it possible to store atoms in the link table, instead of the
concept table.  Since link target is now a property, this connects
naturally with the property storage mechanism.  Most of the changes here
are generalizations and removal of link mapping assumptions.