caos: Implement "exposed behaviour" schema attribute for links
Semantically all schema paths represent sets of nodes.  For practical
reasons, whenever a given set is known to contain not more than one
element, that element (or None, if set is empty) becomes the value of an
attribute representing the link.

This commit adds a mechanism to override this behaviour by explicitly
specifying the "exposed_behaviour" link prototype attribute.  If set to
"first-item", attributes representing the link will always contain the
first element of the set.  If "exposed_behaviour" is "set", attributes
will always contain a corresponding set selector.

The main motivation behind this change is the ability to retain
singleton-like behaviour for non-singular links in the context of
forthcoming schema-level policy-based filtering framework that will
serve as basis for transparent access control and localization.