caos: Remove implicit cast from <Concept> to <Concept>.id in queries
The implicit conversion of entity set references to id references is
dubious and confusing.  The case stems from initial motivation of
allowing equality operators on concept references, but subsequently the
conversion was unnecessarily widely applied.

Since the equality operator case is handled explicitly in
process_binop(), the remaining heuristic-based conversion can be safely
dropped.

As an effect of this change, it is now possible to run aggregate
functions on concept refs, most notably agg:list, which will produce an
array of entities as expected.  Here we go a step further, and filter
out None values from the result, as the default behaviour of backend
array aggregate is to include Nones, and these are often produced by an
associated outer join in the backend query.  In some cases it may be
necessary to retain None values in the resulting array, but supporting
that case is a matter of subsequent commits.