caos.backends.pgsql: Use standalone composite type for concept records
Concepts are extracted from the backend storage as records.  Present
implementation uses concept table as a composite type for concept
records, which can obviously include only one-to-one atoms.

For greater flexibility, create and maintain a standalone composite type
for concept records.  This implementation includes only eagerly-loaded
links into the composite type, so the loading shape is still mostly the
same--singular atoms.

Initially I considered creating composite types for links as well, and
this commit contains some infrastructure to support that, but this does
not seem to be currently necessary, and the complexity price is too high
to do it purely for source metaobject consistency.

It is worth mentioning that all versions of PostgreSQL up to 9.2
disallow self-containment in composite types, thus an attempt to include
a type that is a part of a loop in the schema graph, would result in an
error.  This limitation is a technical one, and a relatively simple
patch for PostgreSQL can be made to allow recursive types.

Another aspect of this change is that tables are now created for all
specialized atomic links with properties.  Previously, link tables
always corresponded to the generic link.