caos: Drop ad-hoc pointer resolution in pyexpr transformer
Historically, links were presented to pyexpr transformer as a special
expr.Link specifier, and not as a specialized link prototype.  This
required pyexpr to perform its own pointer resolution.  The motivation
behind not using link prototypes directly was the intended support for
complex link expressions.  It was found, however, that link expressions
have little practical value to add to link inheritance, and had a very
high price in transformer code complexity.