caos: Do not re-fetch entities with class not matching that of query root
It's more efficient to just do a partial load of an entity and treat
unloaded pointers as expired, than it is to re-fetch every entity
separately right away.  Even more so, it is unlikely that the code
fetching entities through a base class would refer to pointers specific
to children.