utils.functional.plugin: Implement generic registry metaclass for pluggable architectures
Add infrastructure to support pluggable architectures where
interface implementaitons are loosely coupled and the specific
implementations to be used are specified in config.

The base metaclass is not meant to be used directly.  A sub-metaclass
should be used instead as a natural namespace for the specific
universe of plugins used.