caos: Use "C" collation when sorting localizable strings
Ideally, we would use locale-specific collation, based on the language,
but PostgreSQL seems to be lacking the ability to parametrize COLLATE
argument, so we will need to create separate copies of generated query
depending on session context, for which a bit more infrastructure is
required.

Meanwhile, "C" collation is sufficient.