caos.backends.pgsql: Fold property filters into join conditions
Now that filtering does not always originate from query's filter clause,
link property filters must behave in accordance when link path is placed
into a disjunction.  The easiest way to acheve this is to move the SQL
filter out of the WHERE clause and into the join condition.  This _may_
potentially cause plan quality degradation, but further analysis of
pathologic cases is required.