caos: New select() syntax for per-query loading behaviour
Currently, the behaviour of loading cascade is fixed and is determined
in the schema by the "loading" property of pointer prototype.  There are
cases, where the pattern of access is fully known, and for performance
reasons a way to specify a per-query loading cascade is desired.  The
newly gained ability to extract arbitrary structures in queries opens a
possiblity to provide a precise declaration of loading cascade inside
the query.

As such, introduce a new select() syntax.

CaosQL::

    SELECT
        <ConceptRef>\[
        <PropagationSpec>[, <PropagationSpec>...]
    \]

where PropagationSpec is

    {<link_glob>|<link_spec>|<property_glob>|<property_spec>}[\[<PropagationSpec>\]]

link_glob is:

    * [ ( <glob_filter> ) ]

and specifies a list of links, satisfying an optional glob_filter

link_spec is

    [ <link_direction> ] <link_name> [ ( <link_endpoint> ) ]

property_glob is

    @* [ ( <glob_filter> ) ]

and specifies a list of properties, satisfying an optional glob_filter

property_spec is

    <property_name>

glob_filter is

    {any loading | loading=<loading value>}

if glob filter is not specified, it defaults to loading=eager

Example:

    USING
        semantix.caos.tests.testdata.geography_schema,
        semantix.caos.tests.testdata.geography_schema AS geo
    SELECT
        Country[
        people[
            ethnicities
        ],

        geo.divisions[
            *(any loading)
        ],

        geography[
            *,
            area
        ],

        head_of_state[
            *,
            @*,
            @geo.from
        ],

        <native_country(geo.Person)
        ]
    WHERE
        Country.name = 'Germany'

will return a Country entity with "people", "divisions", "geogrpahy" and
"head_of_state" links loaded, in addition to a "native_country" inbound
links from Person entities.  Each link target will be loaded according
to the specified loading cascade declaration.

Python API::

Selectors gained a new method "select()", which is semantically identical
to the extended SELECT clause of CaosQL:

    res = C.select([
            (C.people, [
                C.people.ethnicities
            ]),

            (C.divisions, [
                C.divisions.iter_pointers(loading=None)
            ]),

            (C.geography, [
                C.geography.iter_pointers(),
                C.geography.area
            ]),

            (C.head_of_state, [
                C.head_of_state.iter_pointers(),
                C.head_of_state.as_link().iter_pointers(),
                C.head_of_state.as_link().from_,
            ]),

            C.follow(P.native_country, caos.types.InboundDirection)
         ]) \
         .filter(C.name == 'Germany')

Here, iter_pointers() method serves the role of pointer glob, and
other pointer specifiers are declared using normal expressions.