caos: Allow more than one (source_object, link_class, target_object) triple to exist
Initially, the one of the semantic rules of the Caos schema was that
each (source_object, link_class, target_object) triple must be unique.  This
worked well until the recent introduction of the l10n system that
requires the possibility of multiple link triples that have the same
source and target, but different link properties.

This rework changes the rules so that link objects are now given their
own unique id, and the triple becomes (source_object, link_object,
target_object).

Now, there was a very good reason why the initial restriction was put in
place, and so the original behaviour is kept by default and will be
rewritten as an explicit link constraint in another changeset.