caos: Introduce a notion of static selector
It is now possible to construct and populate a selector with a static
list of values.

This is useful in situations where the protocol requires a proper
selector to be returned, but it is not practical or possible to run the
query, and the data is already available.

Another novelty is the ability to produce "annotated" or "loaded" atom
selectors for cases where it is necessary to, for example, give custom
names to atom values (True/Yes, False/No), etc.  In this case the
selector will be a subclass of LoadedAtomObject and each element will
have a set of attributes accessible through the
__sx_get_atom_attribute__(name) method.