caos: Add access control hooks as an interim mechanism until full AC policy support lands
It is now possible to register callbacks to provide ad-hoc access
control checks and rewrites:

    caos.types.register_access_control_hook(node_cls, 'read'|'write', callback)

'read' callbacks will receive an EntitySet query tree node, which
they should then augment as necessary to achieve the necessary query
result.

'write' callbacks are called from session level and are passed the
session, the entity and the name of the link being modified.  The write
hook should raise an exception should an access violation occur.