caos.caosql: Simplify link inheritance resolution
CaosQL transformer would no longer generate complex disjunctions when
resolving link inheritance, but use the specified link and a generated
virtual target.  There are certainly more places where this
linearization is necessary, but this is a step in the right direction.

Unfortunately, this breaks queries with references to 1-to-1 atoms that
are not present in all possible targets of the link, since the backend
is unable to generate a virtual relation containing the necessary
column.  In practice, this access pattern is very uncommon and usually
signals a badly designed query.