caos: Add rudimentary support for specifying filters in select() specifiers
It is now possible to add a filter to a path expression in select().
This commit adds transformer support as well as exposure for Python API:
just call .filter() on a path expression passed to select().

This is a work in progress: still need to add CaosQL support and
thorough tests.