caos: Complete reimplementation of constraints system
This commit is a culmination of an effort to implement schema
constraints properly as first-class schema objects.

The current implementation, whilst modified and improved a few times, is
essentially the original constraint system, where constraint classes are
hardcoded in protoschema implementation.  This makes it very hard to
define new constraint classes and is also contrary to general Caos
schema philosophy.

With constraints tacked on in such a way, the backend implementation has
also developed quite a few ugly warts in attempts to maintain
consistency during schema changes, especially when inheritance hierarchy
changes.

Hardcoded constraint class implementation also lead to a lack of a
sufficient amount of metadata to correctly implement all constraint
classes using native backend constructs, so several constraint classes
were not enforced by the backend.

The solution is fairly straightforward architecturally.  Make
constraints a first-class schema object: explicitly definable,
inheritable and managed by the delta system as a standalone class.
The constraints would then attach to their subject (known as
ConsistencySubject in protoschema parlour) in a fashion very similar to
how Pointers are attached to Source classes.

In fact, the internal implementation of associated prototype references
(which is a common name for source-pointer-like schema associations)
is now almost completely shared between source-pointer and
subject-constraint protoschema implementations.

In this new implementation constraints are defined similarly to other
top-level schema objects.

For example:

    constraints:
        max:
            expr: subject <= $param
            errmessage: 'Maximum allowed value for {subject} is {param}.'

        length:
            subject: str::len(cast(subject as str))

        maxlength:
            extends: [max, length]
            errmessage: '{subject} must be no longer than {param} characters.'

The above snippet we define "max" as a generic upper-limit constraint.
It dictates that "subject" must not compare larger than "$param".
Here, "subject" is an anchor that will be replaced by the real object
expression when this constraint is applied to a subject class.  "subject"
can also be a function of itself, as demonstrated in the definition of an
abstract "length" constraint.  There, subject is re-defined as length of
subject's character representation.  Finally, by using good old
composition we define "maxlength" as an abstract constraint that limits
the length of character representation of its subject.

Both the main constraint expression and subject definition are CaosQL
expressions.

The constraints defined above are what is called generic constraints,
much like there are generic pointers.  For a constraint to become
useful, it must be assigned to a subject.

For example:

    atoms:
        country_code_t:
	    extends: str
	    constraints:
	        maxlength: 3

In the above example a maxlength constraint is assigned to
country_code_t atom.  In other words, the collection of constraints on
country_code_t atom now contains a specialized constraint derived from a
generic maxlength constraint.  "3" is a value for $param variable in
maxlength constraint expression, which finalizes the constraint
expression:

   str::len(cast(VALUE as str)) <= 3

Here VALUE is a value passed to country_code_t constructor.  This
constraint is also represented as a CHECK constraint on country_code_t
domain in Postgres.

Constraints work similarly when assigned to specialized pointers:

    concepts:
        Country:
            links:
                code:
                    str:
                        constraints:
                            maxlength: 3

Similarly to atom constraints, this dictates that character length of
Country.code must be less or equal to 3.  This translates to a table
constraint in Postgres backend.  Note that, unlike current
implementation, this constraint is defined on a Country.code _linke_,
and not on a magically derived atom Country.code points to.
Automatically derived atoms are not necessary anymore and have been
removed, noticeably simplifying schema operations.

Another significant improvement brought by the new constraint system is
the ability to define arbitrary unique constraints:

    concepts:
        Country:
            links:
                code:
                    str:
                        constraints:
                            unique: str::lower(subject)

The above ensures case-insensitive uniqueness of Country.code, something
that was not possible to define in Caos schema before.

Caos builtins module defines a set of useful constraints that fully implements
current hardcoded versions: maxlength, minlength, regexp, enum, and
unique.  The latter is defined as "not exists (subject)" and differs
from others in that it is a global constraint, i.e it enforces
consistency of a (universal) set of subjects, rather than a single
subject.

Finally, the new constraint framework lays ground for Concept-level
constraints allowing to use arbitrary CaosQL expressions for such
constraints.  There is no support for such constraints in this commit,
but it's straightforward to implement now.

In addition to the above main theme there is a number of significant
refactorings all over Caos to support the new features.

Query backend selecting heuristics has been removed.

The way the decision where to dispatch the query is currently made with
a separate heuristic function, that attempts to determine the scope of
the query.  This approach is inherently flawed and unreliable.  Instead,
push the decision making into the Python transformer.  If, for any
reason, the query cannot be processed it would raise
NotImplementedError, indicating that a more capable backend should take
over.

Path variables and anchors are no longer mixed.

Historically path variables were called anchors internally, and when a
concept of real anchors was also added, it used the same internal
nomenclature and structures.  Since anchors are now relied upon much
more heavily, the trasformers got confused by path variables being
present in the same place.  Fix this by splitting out path variables
and tracking them separately from anchors.  An example of an anchor is
"subject" in constraint expressions.

Pointer defaults are no longer merged during inheritance merge.

This is an unused feature that has been broken by a more stringent
validation of default expressions, since there is no way to remove or
alter an expression in parents, and it is possible for those expressions
to generate invalid output in the context of a child.

It is now possible to evaluate detached selectors if they need not
be processed by the backend data, i.e. simple expressions operating on
data that is already present.